requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
redis==5.0.1
cryptography==42.0.5
Flask-Limiter==3.5.0
Werkzeug>=3.0.0
psycopg2-binary==2.9.9
